+++ packages/enraiged/src/Users/Models/Scopes/Reportable.php
@@ -21,6 +21,7 @@ trait Reportable
         (object) $this
             ->addSelect($columns)
             ->joinProfiles()
+            ->joinCountries()
             ->joinRoles()
             ->whereVisible();
 
@@ -36,7 +37,23 @@ trait Reportable
             ? collect($columns)->join(',')
             : $columns;
 
-        $this->reportable->addSelect($columns);
+        $this->reportable
+            ->addSelect($columns);
+
+        return $this;
+    }
+
+    /**
+     *  @return self
+     */
+    private function joinCountries(): self
+    {
+        $this->reportable
+            ->leftJoin('addresses', fn ($join)
+                => $join
+                    ->on('addresses.addressable_id', 'profiles.id')
+                    ->where('addresses.addressable_type', '=', Enraiged\Profiles\Models\Profile::class))
+            ->leftJoin('countries', 'countries.id', '=', 'addresses.country_id');
 
         return $this;
     }
@@ -46,7 +63,8 @@ trait Reportable
      */
     private function joinProfiles(): self
     {
-        $this->reportable->join('profiles', 'profiles.id', '=', 'users.profile_id');
+        $this->reportable
+            ->join('profiles', 'profiles.id', '=', 'users.profile_id');
 
         return $this;
     }
@@ -57,9 +75,11 @@ trait Reportable
     private function joinRoles(): self
     {
         if (config('enraiged.auth.force_lowest_role')) {
-            $this->reportable->join('roles', 'roles.id', '=', 'users.role_id');
+            $this->reportable
+                ->join('roles', 'roles.id', '=', 'users.role_id');
         } else {
-            $this->reportable->leftJoin('roles', 'roles.id', '=', 'users.role_id');
+            $this->reportable
+                ->leftJoin('roles', 'roles.id', '=', 'users.role_id');
         }
 
         return $this;